+++ src/frontend/my-app/src/business/hooks/CppAnalyzer.ts
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { callAnalysisAPI } from "../services/AnalyzeService";
-import  APIError  from "../models/APIError";
+import APIError from "../models/APIError";
 
 interface AnalysisResult {
     compilationStatus: 'success' | 'error' | 'pending';
@@ -14,13 +14,14 @@ export const useCppAnalyzer = () => {
     const [problemDescription, setProblemDescription] = useState('');
     const [isAnalyzing, setIsAnalyzing] = useState(false);
     const [analysisResult, setAnalysisResult] = useState<AnalysisResult | null>(null);
-    const [error, setError] = useState('');
+    const [error, setError] = useState<string | null>(null);
+
 
     const handleFileUpload = (event: React.ChangeEvent<HTMLInputElement>) => {
         const uploadedFile = event.target.files?.[0];
         if (uploadedFile && uploadedFile.name.endsWith('.cpp')) {
             setSolutionFile(uploadedFile);
-            setError('');
+            setError(null);
             const reader = new FileReader();
             reader.onload = (e) => setFileContent(e.target?.result as string);
             reader.readAsText(uploadedFile);
@@ -34,6 +35,8 @@ export const useCppAnalyzer = () => {
         return new File([blob], 'problem.txt', { type: 'text/plain' });
     };
 
+    const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+
     const analyzeCode = async () => {
         if (!solutionFile || !fileContent || !problemDescription.trim()) {
             setError('Vui lòng nhập đầy đủ mô tả và chọn file C++');
@@ -42,14 +45,15 @@ export const useCppAnalyzer = () => {
 
         setIsAnalyzing(true);
         setAnalysisResult(null);
-        setError('');
+        setError(null);
 
         try {
             const problemFile = problemTextToFile(problemDescription);
+            await sleep(300);
             const apiResponse = await callAnalysisAPI(solutionFile, problemFile);
 
             setAnalysisResult({
-                compilationStatus: 'success',
+                compilationStatus: apiResponse.status === 'success' ? 'success' : 'error',
                 compilationErrors: apiResponse.data?.compile_output || '',
                 llmReview: apiResponse.data?.llm_feedback || '',
             });
